rag-pipeline/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
chromadb
openai
langchain
langchain-text-splitters
pymupdf
python-docx
python-dotenv
pytest
httpx
